pyproject.toml
[project]
name = "repo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "cloudflared>=1.0.0.2",
    "fastapi>=0.120.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pypdf>=6.1.3",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
    "streamlit>=1.50.0",
    "uvicorn[standard]>=0.38.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.8.0",
]
